Migrate the register items test to a request test
Controller tests are deprecated
